Use system separator and indicator
